% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_determine_tmb.R
\name{determine_mutational_burden}
\alias{determine_mutational_burden}
\title{Determine mutational burden based on GRCm39.}
\usage{
determine_mutational_burden(x)
}
\arguments{
\item{x}{(SimpleVRangesList): VRangesList containing the mutations}
}
\value{
(tibble): Tibble with information about the mutational burden.
}
\description{
Determine mutational burden based on GRCm39.
}
