% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_determine_titv.R
\name{determine_ti_tv}
\alias{determine_ti_tv}
\title{Determine Ti/Tv ratio.}
\usage{
determine_ti_tv(m_sbs)
}
\arguments{
\item{m_sbs}{(matrix): Mutational SBS matrix.}
}
\value{
(tibble): Tibble with information about the Ti/Tv.
}
\description{
Determine Ti/Tv ratio.
}
