% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_import_samples.R
\name{import_samples}
\alias{import_samples}
\title{Import WGS and WTS data from the snakemake-lesionsegration workflow.}
\usage{
import_samples(metadata, workflow_dir, gtf = NULL)
}
\arguments{
\item{metadata}{(tibble): Metadata of the samples.}

\item{workflow_dir}{(character): Path to the workflow output directory.}

\item{gtf}{(tibble): Tibble to imported/filtered GTF.}
}
\value{
(list): List of 96-context matrices (SBS, InDel, DBS).
}
\description{
Import WGS and WTS data from the snakemake-lesionsegration workflow.
}
