% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_import_vcf.R
\name{import_vcf}
\alias{import_vcf}
\title{Importing VCF files.}
\usage{
import_vcf(path_vcf)
}
\arguments{
\item{x}{(character): Path to one VCF file(s).}
}
\value{
(tibble): Tibble containing the (filtered) somatic variants.
and strain-specific tagging.
}
\description{
Imports VCF files from the snakemake-lesionsegregation workflow.
Applies several filtering steps and determines the likely strain-origin of the somatic variants.
}
\examples{
1 + 1
}
