% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_import_vcf_mutect.R
\name{import_vcf_mutect}
\alias{import_vcf_mutect}
\title{Importing VCF files.}
\usage{
import_vcf_mutect(x)
}
\arguments{
\item{x}{(character): Path to one or multiple VCF file(s).}
}
\value{
(VRanges): VRanges object containing the (filtered) somatic variants
and strain-specific tagging.
}
\description{
Imports VCF files from the snakemake-lesionsegregation workflow.
Applies several filtering steps and determines the likely strain-origin of the somatic variants.
}
\examples{
1 + 1
}
