% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_read_flagstats.R
\name{read_flagstats}
\alias{read_flagstats}
\title{Import and read flagstats}
\usage{
read_flagstats(x)
}
\arguments{
\item{x}{(character): Paths to one or multiple flagstats file.}
}
\value{
(tibble): Tibble with columns variable, value, and sample.
}
\description{
Import and read flagstats
}
