% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_read_haplolog.R
\name{read_haplotag_log}
\alias{read_haplotag_log}
\title{Import and read haplotag log.}
\usage{
read_haplotag_log(x)
}
\arguments{
\item{x}{(character): Paths to one or multiple flagstats file.}
}
\value{
(tibble): Tibble with columns variable, value, and sample.
}
\description{
Determine the no. of tagged reads and used variants during haplotagging.
}
