% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_read_snpsplit_yaml.R
\name{read_snpsplit_yaml}
\alias{read_snpsplit_yaml}
\title{Import and clean SNPsplit reporting yamls}
\usage{
read_snpsplit_yaml(x)
}
\arguments{
\item{x}{(character): Path to flagstats file.}
}
\value{
(tibble): Tibble containing the SNPsplit reporting yaml.
}
\description{
Import and clean SNPsplit reporting yamls
}
