% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_run_dnds.R
\name{run_dnds}
\alias{run_dnds}
\title{Perform dN/dS analysis.}
\usage{
run_dnds(data_muts, path_db = "~/Downloads/refCDS_ENSEMBLv109_GRCm39.rda")
}
\arguments{
\item{data_muts}{(SimpleVRangesList): VRangesList containing the mutations of all samples.}

\item{path_db}{(character): Path to the dN/dS database.}
}
\value{
(list): Tibble with
}
\description{
Perform dN/dS analysis.
}
\details{
The database can be generated using the following code:

ENSEMBLv109_GRCm39 <- readr::read_tsv("~/Downloads/mart_export.txt") %>%
    dplyr::filter(!grepl("_", `Chromosome/scaffold name`)) %>%
    dplyr::filter(`CDS Length` %% 3 == 0) %>%
    dplyr::filter(!is.na(`Genomic coding start`)) %>%
    dplyr::filter(!grepl("^AP00|^RP11-", `Gene name`)) %>%
    dplyr::filter(!grepl("\\.", `Gene name`)) %>%
   dplyr::select(
        "gene.id" = `Gene stable ID`,
        "gene.name" = `Gene name`,
        "cds.id" = `Protein stable ID`,
        "chr" = `Chromosome/scaffold name`,
        "chr.coding.start" = `Genomic coding start`,
        "chr.coding.end" = `Genomic coding end`,
        "cds.start" = `CDS start`,
        "cds.end" = `CDS end`,
        "length" = `CDS Length`,
        "strand" = Strand
    ) %>%
dplyr::mutate(chr = paste0("chr", chr))
pathCDS = '~/test/mart_export_filtered.txt'
pathFasta = '/omics/groups/OE0538/internal/projects/sharedData/GRCm39/genome/GRCm39.primary_assembly.genome.fa'
dndscv::buildref(cdsfile = pathCDS, genomefile = pathFasta, outfile = '~/test/refCDS_ENSEMBLv109_GRCm39.rda', excludechrs='chrMT', useids = TRUE)
}
