% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_generate_96muts.R
\name{generate_mutmatrices_96}
\alias{generate_mutmatrices_96}
\title{Generate 96-context matrix of somatic variants.}
\usage{
generate_mutmatrices_96(x)
}
\arguments{
\item{x}{(VRanges): VRanges containing the mutations of the samples.}
}
\value{
(list): List of 96-context tibbles (SBS and InDel).
}
\description{
Generate 96-context matrix of somatic variants.
}
